model_tests/app_project/requirements.txt
Flask
Flask-Cors
opencv-python
numpy
boto3
botocore
cryptography
snowflake-connector-python
deepface
face_recognition
